+++ node/tests/api/template/render.test.ts
@@ -1,0 +1,66 @@
+/**
+ * Copyright (C) 2023 SiYuan Community
+ * 
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
+ * 
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ * 
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+import {
+    describe,
+    test,
+    expect,
+} from "vitest";
+
+import client from "~/tests/utils/client";
+import { testKernelAPI } from "~/tests/utils/test";
+import { SchemaJSON } from "~/tests/utils/schema";
+import constants from "~/tests/constants";
+
+import render from "@/types/kernel/api/template/render";
+
+const pathname = client.Client.api.template.render.pathname;
+
+describe(pathname, async () => {
+    const schema_payload = new SchemaJSON(SchemaJSON.resolvePayloadSchemaPath(pathname));
+    const schema_response = new SchemaJSON(SchemaJSON.resolveResponseSchemaPath(pathname));
+    await schema_payload.loadSchemaFile();
+    await schema_response.loadSchemaFile();
+    const validate_payload = schema_payload.constructValidateFuction();
+    const validate_response = schema_response.constructValidateFuction();
+
+    testKernelAPI<render.IPayload, render.IResponse>({
+        name: "main",
+        payload: {
+            data: {
+                id: "20200812220555-lj3enxa", // 思源笔记用户指南/请从这里开始
+                path: `${process.env.VITE_SIYUAN_WORKSPACE_PATH}/temp/convert/test/render/test.md`,
+            },
+            validate: validate_payload,
+            test: async () => {
+                await client.client.putFile({
+                    path: "/temp/convert/test/render/test.md",
+                    file: constants.TEST_TEMPLATE_CONTENT,
+                });
+            },
+        },
+        request: (payload) => client.client.render(payload!),
+        response: {
+            validate: validate_response,
+            test: response => {
+                test("test the result of rendering template file", () => {
+                    expect(response.data.path).toMatch(`${process.env.VITE_SIYUAN_WORKSPACE_PATH}/temp/convert/test/render/test.md`);
+                });
+            },
+        },
+    });
+});